+++ darts/models/forecasting/patchtst_fm_model.py
@@ -35,6 +35,7 @@ from darts.models.forecasting.pl_forecasting_module import (
 )
 from darts.utils.data.torch_datasets.utils import (
     InputChunkLength,
+    ModuleStage,
     PLModuleInput,
     PLModuleOutput,
     TorchTrainingSample,
@@ -338,7 +339,7 @@ class _PatchTSTFMModule(PLForecastingModule):
         # during training, output all pre-trained quantiles for loss
         # during prediction, output only user-specified quantiles
         # -> (B, T, C, N)
-        if self.training:
+        if x_in.stage is ModuleStage.TRAIN:
             q_forecast = q_forecast[:, :, :, self._finetuning_quantile_indices]
         else:
             q_forecast = q_forecast[:, :, :, self.user_quantile_indices]