+++ darts/models/forecasting/timesfm2p5_model.py
@@ -38,6 +38,7 @@ from darts.models.forecasting.pl_forecasting_module import (
 )
 from darts.utils.data.torch_datasets.utils import (
     InputChunkLength,
+    ModuleStage,
     PLModuleInput,
     PLModuleOutput,
     TorchTrainingSample,
@@ -314,7 +315,7 @@ class _TimesFM2p5Module(PLForecastingModule):
 
         # during training (fine-tuning), output all pre-trained quantiles for loss;
         # during prediction, output only user-specified quantiles
-        if self.training:
+        if x_in.stage is ModuleStage.TRAIN:
             renormed_outputs = renormed_outputs[
                 :, :, :, self._finetuning_quantile_indices
             ]